Refactor code and remove unused files
